Rename zfs_sb_t -> zfsvfs_t
The use of zfs_sb_t instead of zfsvfs_t results in unnecessary
conflicts with the upstream source.  Change all instances of
zfs_sb_t to zfsvfs_t and update the variables names.

When possible the code was updated to match the OpenZFS
code.

One way to further reduce differences would be to push the
ZTOZSB and related macros in to OpenZFS.  This would allow
each platform to implement the right thing for their needs.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>