Add -M (mmp_test) option to ztest
When using ztest to generate activity, for automated tests of the
MMP function, some test functions interfere with the test.

The currently known function that is problematic is ztest_rename,
as the test attempts to import the pool by name.  Others are
expected to be identified.

Add a new ztest function, "-M", to alter ztest behavior to prevent
this.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>